add modal withdrawl except css setting and currentPassword error issue remains

+++ frontend/src/app/dashboard/profile/profile.component.ts
@@ -24,10 +24,10 @@ export class ProfileComponent implements OnInit {
 
   public currentUser: User;
   public passwordForm: FormGroup;
-  public password: AbstractControl;
-  public passwordConfirm: AbstractControl;
-  public currPassword: string;
-  public showDialog: boolean;
+  public modalPasswordForm: FormGroup;
+  public currentPassword: AbstractControl;
+  public newPassword: AbstractControl;
+  public newPasswordConfirm: AbstractControl;
   private pastRooms: Room[];
 
 
@@ -40,50 +40,53 @@ export class ProfileComponent implements OnInit {
   }
 
   ngOnInit() {
-    this.showDialog = true;
     this.accountService.getUserDetail()
       .then(user => {
         this.currentUser = user;
+
+        // password change form
         this.passwordForm = this.formBuilder.group({
-          'password' : [ '', Validators.required ],
-          'passwordConfirm' : [ '', Validators.required ]
+          'currentPassword' : [ '', ],
+          'newPassword' : [ '', Validators.required ],
+          'newPasswordConfirm' : [ '', Validators.required ]
         }, {
           validator : PasswordValidator.matchForm
         });
-        this.password = this.passwordForm.controls[ 'password' ];
-        this.passwordConfirm = this.passwordForm.controls[ 'passwordConfirm' ];
+        this.currentPassword = this.passwordForm.controls[ 'currentPassword' ];
+        this.newPassword = this.passwordForm.controls[ 'newPassword' ];
+        this.newPasswordConfirm = this.passwordForm.controls[ 'newPasswordConfirm' ];
+
+        this.modalPasswordForm = this.formBuilder.group({
+          'passwordCheck' : [ '', ]
+        });
+
+        this.meetService.getRoomsJoinedPast()
+          .then(rooms => {
+            this.pastRooms = rooms;
+          });
       });
-    this.meetService.getRoomsJoinedPast()
-      .then(rooms => {
-        this.pastRooms = rooms;
-        console.log(rooms);
-      });
+
   }
 
   changePassword() {
-    this.currentUser.password = this.password.value;
-    this.accountService.putUser(this.currentUser)
-      .then(isSuccessToPut => {
-        if (isSuccessToPut) {
-          this.router.navigate([ 'dashboard' ]);
+    this.accountService.checkPassword(this.currentPassword.value)
+      .then(checked => {
+        if (checked) {
+          this.currentUser.password = this.newPassword.value;
+          this.accountService.putUser(this.currentUser)
+            .then(isSuccessToPut => {
+              if (isSuccessToPut) {
+                this.router.navigate([ 'dashboard' ]);
+              }
+            });
+        } else {
+          this.currentPassword.setErrors({ invalidCurrentPassword : true });
         }
       });
   }
 
   goBack(): void {
     this.router.navigate([ 'dashboard' ]);
-  }
-
-  checkCurrentPassword(): void {
-
-    // for null value
-    if (!this.currPassword)
-      return;
-    this.accountService.checkPassword(this.currPassword)
-      .then(res => {
-        if (res)
-          this.showDialog = false;
-      });
   }
 
   showUserInfo(): void {
@@ -99,20 +102,26 @@ export class ProfileComponent implements OnInit {
   public open(dynamicContent: string = 'Example') {
     const config = new TemplateModalConfig<IContext, string, string>(this.modalTemplate);
 
-    config.closeResult = 'closed!';
+    config.size = 'mini';
     config.context = { data : dynamicContent };
 
     this.modalService
       .open(config)
-      .onApprove(result => {
-        this.accountService.deleteUser()
-          .then(deleteSuccess => {
-            console.log(deleteSuccess);
-            if (deleteSuccess) {
-              this.router.navigate([ 'login' ]);
+      .onApprove(() => {
+        this.accountService.checkPassword(this.modalPasswordForm.controls[ 'passwordCheck' ].value)
+          .then(checked => {
+            if (checked) {
+              this.accountService.deleteUser()
+                .then(deleteSucces => {
+                  if (deleteSucces) {
+                    this.router.navigate([ 'login' ]);
+                  }
+                });
+            } else {
+              this.modalPasswordForm.controls[ 'passwordCheck' ].setErrors({ passwordCheckFailed : true });
             }
           });
       })
-      .onDeny(result => { /* deny callback */});
+      .onDeny(() => { this.modalPasswordForm.controls[ 'passwordCheck' ].setValue('');});
   }
 }
